+++ tests/test_model.py
@@ -349,7 +349,7 @@ class TestModel(TestCase):
         self.assertOutput('')
 
     def test_standard_methods(self):
-        for method in 'write create copy unlink'.split():
+        for method in 'write', 'copy', 'unlink':
             self.test_method(method)
 
         self.test_method('perm_read', single_id=False)